started porting merging fixed joints into a separate file
